Organize redux store
