+++ src/components/App.connect.tsx
@@ -1,23 +1,18 @@
-import { ActionType } from 'typesafe-actions';
 import { Dispatch } from 'redux';
 import { connect } from 'react-redux';
 
-import * as actions from "../actions";
-
-import { RootState } from "../reducers";
+import { ActionsType, RootStateType } from "../shared/store";
 
 import App from "./App";
-
-type Action = ActionType<typeof actions>;
 
 interface OwnProps {
 }
 
-const mapStateToProps = (state: RootState) => ({
+const mapStateToProps = (state: RootStateType) => ({
   loading: !state.map.ready,
 });
 
-const mapDispatchToProps = (dispatch: Dispatch<Action>, props: OwnProps) => ({
+const mapDispatchToProps = (dispatch: Dispatch<ActionsType>, props: OwnProps) => ({
 });
 
 export default connect(mapStateToProps, mapDispatchToProps)(App);
