+++ src/components/Map.connect.tsx
@@ -1,21 +1,16 @@
-import { ActionType } from 'typesafe-actions';
 import { Dispatch, bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 
-import * as actions from "../actions";
-
-import { RootState } from "../reducers";
-
-type Action = ActionType<typeof actions>;
+import { ActionsType, RootStateType, actions } from "../shared/store";
 
 import Map from "./Map";
 
 interface OwnProps {
 }
 
-const mapStateToProps = (state: RootState) => ({});
+const mapStateToProps = (state: RootStateType) => ({});
 
-const mapDispatchToProps = (dispatch: Dispatch<Action>, props: OwnProps) => bindActionCreators({
+const mapDispatchToProps = (dispatch: Dispatch<ActionsType>, props: OwnProps) => bindActionCreators({
   getWeather: (lat: number, lng: number) => actions.weatherGetAction(lat, lng),
   mapReady: () => actions.mapReadyAction(),
 }, dispatch);
