+++ src/shared/store/constants/index.ts
