+++ src/shared/store/reducers/MapReducer.ts
